completed drivetrain subsystem
