+++ Competition/src/main/cpp/Robot.cpp
@@ -28,6 +28,7 @@ void Robot::RobotPeriodic() { frc2::CommandScheduler::GetInstance().Run(); }
  * robot is disabled.
  */
 void Robot::DisabledInit() {
+  m_container.m_drivetrain.setDefaultState();
   m_container.m_shooter.setDefaultState();
   m_container.m_intake.setDefaultState();
   m_container.m_hopper.setDefaultState();
