+++ Competition/src/main/include/Constants.h
@@ -58,8 +58,9 @@ namespace DriveConstants {
     constexpr static double kAllError = 0;
     constexpr static double kDeadband = 0.05;
 
-    constexpr static double kDeadbandY = 0.05;
     constexpr static double kDeadbandX = 0.05;
+    constexpr static double kDeadbandY = 0.05;
+    constexpr static double kDeadbandTrigger = 0.05;
     constexpr static double kDriveMultiplierX = 0.5;
     constexpr static double kDriveMultiplierY = 1;
     constexpr static double kDriveOffset = 1;
