+++ Competition/src/main/include/Drivetrain.h
@@ -1,124 +1,101 @@
-/*----------------------------------------------------------------------------*/
-/* Copyright (c) 2017-2018 FIRST. All Rights Reserved.                        */
-/* Open Source Software - may be modified and shared by FRC teams. The code   */
-/* must be accompanied by the FIRST BSD license file in the root directory of */
-/* the project.                                                               */
-/*----------------------------------------------------------------------------*/
-
 #pragma once
 
-#include <frc2/command/Subsystem.h>
-#include <frc/SpeedControllerGroup.h>
+#include "ValorSubsystem.h"
+#include "Constants.h"
+
+#include <frc/XboxController.h>
+#include <frc/geometry/Pose2d.h>
+#include <frc/kinematics/DifferentialDriveOdometry.h>
+#include <units/units.h>
+#include <frc/kinematics/DifferentialDriveWheelSpeeds.h>
+#include <frc/kinematics/DifferentialDriveKinematics.h>
+#include <adi/ADIS16448_IMU.h>
+#include <frc/controller/SimpleMotorFeedforward.h>
+#include <frc/trajectory/TrajectoryConfig.h>
+#include <frc/trajectory/constraint/DifferentialDriveVoltageConstraint.h>
 #include <rev/CANSparkMax.h>
 #include <rev/CANEncoder.h>
-
-#include <frc/kinematics/DifferentialDriveOdometry.h>
-#include <frc/geometry/Pose2d.h>
-#include <units/units.h>
-#include <frc/geometry/Rotation2d.h>
-#include <frc/kinematics/DifferentialDriveWheelSpeeds.h>
-#include <frc/controller/SimpleMotorFeedforward.h>
-#include <frc/kinematics/DifferentialDriveKinematics.h>
-#include <frc/trajectory/constraint/DifferentialDriveVoltageConstraint.h>
-#include <frc/trajectory/Trajectory.h>
-#include <frc/trajectory/TrajectoryGenerator.h>
-#include <adi/ADIS16448_IMU.h>
+#include <rev/CANPIDController.h>
 
 #include <networktables/NetworkTableEntry.h>
 #include <networktables/NetworkTableInstance.h>
 #include <networktables/NetworkTable.h>
-#include <frc/livewindow/LiveWindow.h>
-#include <frc/smartdashboard/SmartDashboard.h>
 
-#include "Constants.h"
+class Drivetrain : public ValorSubsystem {
+    public:
+        Drivetrain();
 
-#ifndef DRIVETRAIN_H
-#define DRIVETRAIN_H
+        void init();
+        void setController(frc::XboxController*);
 
-class Drivetrain : public frc2::Subsystem {
- private:
-  rev::CANSparkMax m_leftDriveLead;
-  rev::CANSparkMax m_leftDriveFollowA;
-  rev::CANSparkMax m_leftDriveFollowB;
-  rev::CANSparkMax m_rightDriveLead;
-  rev::CANSparkMax m_rightDriveFollowA;
-  rev::CANSparkMax m_rightDriveFollowB;
+        void setDefaultState();
+        void assessInputs();
+        void assignOutputs();
 
-  frc::ADIS16448_IMU imu{};
+        void resetState();
 
-  rev::CANPIDController m_leftPIDController = m_leftDriveLead.GetPIDController();
-  rev::CANPIDController m_rightPIDController = m_rightDriveLead.GetPIDController();
+        void setPower(double power);
 
-  rev::CANEncoder m_leftEncoder = m_leftDriveLead.GetEncoder();
-  rev::CANEncoder m_rightEncoder = m_rightDriveLead.GetEncoder();
+        void resetEncoders();
+        void resetOdometry(frc::Pose2d pose);
+        void resetIMU();
 
-  frc::DifferentialDriveOdometry m_odometry;
+        units::meter_t getLeftDistance();
+        units::meter_t getRightDistance();
+        double getHeading();
+        double getTurnRate();
+        frc::Pose2d getPose();
+        frc::DifferentialDriveWheelSpeeds getWheelSpeeds();
+        void tankDriveVolts(units::volt_t leftVolts, units::volt_t rightVolts);
 
-  double directionY;
-  double straightValue;
-  double straightTarget;
-  double directionX;
-  double turnValue;
-  double turnTarget;
+        enum DrivetrainState {
+            DISABLED,
+            AUTO,
+            MANUAL
+        };
 
-  int limelight_state;
+        struct x {
+            DrivetrainState drivetrainState;
 
- public:
-  Drivetrain();
+            double leftTrigger;
+            double rightTrigger;
+            double leftJoystickX;
+            bool Ybutton;
+            bool rightBumper;
 
-  // global method to return instance of drive subsystem
-  static Drivetrain& GetInstance();
+            double directionX;
+            double directionY;
+            double boostMultiplier;
+            
+            double straightTarget;
+            double turnTarget;
+            double currentLeftTarget;
+            double currentRightTarget;
+            int limelightState;
+        } state;
 
-  // loop method to always update odometry and state of subsystem
-  void Periodic() override;
+        frc::DifferentialDriveKinematics kDriveKinematics;
+        frc::SimpleMotorFeedforward<units::meters> kSimpleMotorFeedforward;
+        frc::TrajectoryConfig kTrajectoryConfigF;
+        frc::TrajectoryConfig kTrajectoryConfigR;
+        frc::DifferentialDriveVoltageConstraint kDifferentialDriveVoltageConstraint;
 
-  // initializes drive motors to proper configurations
-  void InitDrives(rev::CANSparkMax::IdleMode idleMode);
+    private:
+        rev::CANSparkMax leftDriveLead;
+        rev::CANSparkMax leftDriveFollowA;
+        rev::CANSparkMax leftDriveFollowB;
+        rev::CANSparkMax rightDriveLead;
+        rev::CANSparkMax rightDriveFollowA;
+        rev::CANSparkMax rightDriveFollowB;
 
-  // resets encoder values
-  void ResetEncoders();
-  // resets odometry to given pose
-  void ResetOdometry(frc::Pose2d pose);
-  // resets imu sensor
-  void ResetIMU();
-  // return average of meters traveled left and right motors
-  double GetEncAvgDistance();
+        rev::CANPIDController leftPIDController = leftDriveLead.GetPIDController();
+        rev::CANPIDController rightPIDController = rightDriveLead.GetPIDController();
+        rev::CANEncoder leftCANEncoder = leftDriveLead.GetEncoder();
+        rev::CANEncoder rightCANEncoder = rightDriveLead.GetEncoder();
 
-  // returns meters traveled by left motors
-  units::meter_t GetLeftDistance();
-  // returns meters traveled by right motors
-  units::meter_t GetRightDistance();
+        frc::ADIS16448_IMU imu{};
 
-  // returns imu heading in degrees
-  double GetHeading();
-  // return angular velocity of drivetrain
-  double GetTurnRate();
+        frc::DifferentialDriveOdometry odometry;
 
-  // returns pose of odometry (x, y, heading)
-  frc::Pose2d GetPose();
-  frc::DifferentialDriveWheelSpeeds GetWheelSpeeds();
-
-  // method to supply voltage to left and right motors individually
-  void TankDriveVolts(units::volt_t leftVolts, units::volt_t rightVolts);
-  // tele-op rocket league controls drive method
-  void RocketLeagueDrive(double straightInput, double reverseInput, double turnInput, bool limelightInput);
-
-  // sets boostMultiplier to 0.5 or 1 for drivetrain speed
-  void SetMultiplier(double multiplier);
-
-  // stops all drive motors
-  void Stop();
-
-  double boostMultiplier;
-
-  frc::DifferentialDriveKinematics kDriveKinematics;
-  frc::SimpleMotorFeedforward<units::meters> kSimpleMotorFeedforward;
-
-  // need forward and reverse trajectory configs for when robot moves in spline forward/reverse
-  frc::TrajectoryConfig kTrajectoryConfigF;
-  frc::TrajectoryConfig kTrajectoryConfigR;
-  frc::DifferentialDriveVoltageConstraint kDifferentialDriveVoltageConstraint;
-
+        frc::XboxController* driverController;
 };
-
-#endif
